requirements.txt
azure-identity
azure-storage-blob
azure-mgmt-resource
azure-mgmt-storage